Few clearups to avoid short pulse of PWM on startup, changes related to field weakenning and settings, function to dynamically monitor flux linkage

+++ MESC_Common/Inc/MESCfoc.h
@@ -125,7 +125,7 @@ typedef struct {
   float Vd_injectionV;
   float Vq_injectionV;
   uint32_t FLrun, VFLrun;
-  uint16_t angle_error;
+  int angle_error;
   float Ldq_now[2];
   float Ldq_now_dboost[2];
